# src/orders/api/openapi.yaml
openapi: 3.1.0
info:
  title: Order Service API
  version: 1.0.0
  description: |
    This API provides order management functionality.

    **Global Middlewares:**
    - Logging: All requests are logged with request/response details.
    
    **External Service Dependencies:**
    - Product Service: Used for product information and pricing (via PRODUCT_SERVICE_URL)
    - Loyalty Service: Used for managing customer loyalty points (via LOYALTY_SERVICE_URL)
contact:
  name: Bitovi Support
  email: support@bitovi.com

servers:
  - url: http://localhost:8080
    description: Local development server

security:
  - {}

paths:
  /health:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      responses:
        '200':
          description: Server is healthy
  
  /orders:
    get:
      summary: Retrieve a list of orders
      description: |
        Retrieves a list of all orders in the system

        **Middlewares applied:**
        - Authentication required (admin role)
      operationId: listOrders
      tags:
        - Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved list of orders
          content:
            application/json:
              schema:
                type: object
                required:
                  - orders
                  - total
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
                    description: Total number of orders available
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new order
      description: |
        Creates a new order in the system

        **Middlewares applied:**
        - Authentication required (admin role)
      operationId: createOrder
      tags:
        - Orders
      security:
        - bearerAuth: []
      requestBody:
        description: Order details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - products
              properties:
                userId:
                  type: string
                  format: uuid
                  description: Unique identifier for the user placing the order
                products:
                  type: array
                  description: List of products in the order with their quantities
                  minItems: 1
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                    properties:
                      productId:
                        type: string
                        format: uuid
                        description: Unique identifier for the product
                      quantity:
                        type: integer
                        description: Quantity of the product ordered
                        minimum: 1
      responses:
        '201':
          description: Successfully created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /orders/{orderId}:
    get:
      summary: Get order by ID
      description: |
        Retrieves a single order by its unique identifier

        **Middlewares applied:**
        - Authentication required (admin role)
      operationId: getOrderById
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          description: Unique identifier of the order
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid order ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update an existing order
      description: |
        Updates products in an existing PENDING order. Only pending orders can be updated.

        **Middlewares applied:**
        - Authentication required (admin role)
      operationId: updateOrder
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          description: Unique identifier of the order to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Updated order details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - products
              properties:
                products:
                  type: array
                  description: List of products to add/update/remove in the order
                  minItems: 1
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                    properties:
                      productId:
                        type: string
                        format: uuid
                        description: Unique identifier for the product
                      quantity:
                        type: integer
                        description: |
                          Quantity change for the product:
                          - Positive: Add quantity to existing or create new product line
                          - Negative: Subtract quantity from existing (removes if result <= 0)
                          - Zero: No change
      responses:
        '200':
          description: Successfully updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order data or order cannot be updated (not pending)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /orders/{orderId}/submit:
    post:
      summary: Cancel or submit an order
      description: |
        Cancels or submits an existing order. Submitting changes status to PROCESSING and triggers loyalty-service accrual.

        **Middlewares applied:**
        - Authentication required (admin role)
      operationId: cancelOrSubmitOrder
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          description: Unique identifier of the order to cancel or submit
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Action to perform on the order
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  description: Action to perform on the order
                  enum:
                    - CANCEL
                    - SUBMIT
      responses:
        '200':
          description: Successfully performed action on order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid action or order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication. Include the token in the Authorization header as "Bearer {token}"
  
  schemas:
    Order:
      type: object
      required:
        - id
        - userId
        - products
        - totalPrice
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the order
        userId:
          type: string
          format: uuid
          description: Unique identifier for the user placing the order
        products:
          type: array
          description: List of products in the order with their quantities
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: string
                format: uuid
                description: Unique identifier for the product
              quantity:
                type: integer
                description: Quantity of the product ordered
                minimum: 1
        totalPrice:
          type: number
          format: float
          description: Total price for the order
          minimum: 0
        accruedLoyaltyPoints:
          type: integer
          description: Loyalty points accrued from this order (calculated by loyalty-service)
          minimum: 0
        orderDate:
          type: string
          format: date-time
          description: Timestamp when the order was placed
        status:
          type: string
          description: Current status of the order
          enum:
            - PENDING
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELED
  
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: "ORDER_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "The requested order could not be found"
        details:
          type: string
          description: Additional error details